delete and new title
