+++ frontend/src/lib/lemon-ui/LemonTree/LemonTree.tsx
@@ -318,13 +318,7 @@ const LemonTreeNode = forwardRef<HTMLDivElement, LemonTreeNodeProps>(
                             data-id={item.id}
                             // When dragging, don't allow links to be clicked,
                             // without this drag end would fire this href causing a reload
-                            to={
-                                item.disabledReason || isEmptyFolder
-                                    ? '#'
-                                    : typeof item.record?.href === 'function'
-                                    ? item.record.href()
-                                    : item.record?.href || '#'
-                            }
+                            to={item.disabledReason || isEmptyFolder ? '#' : item.record?.href || '#'}
                             onClick={(e) => {
                                 if (item.disabledReason) {
                                     e.preventDefault()
